requirements.txt
Flask==2.2.2
pymongo==4.1.1
bcrypt==4.0.1
Werkzeug==2.2.2
pycryptodome == 3.15.0
pytz == 2021.1
rich==13.4.2